refactor: GraphQL server initialization
